requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
mysql-connector-python
tzdata
python-dateutil
pytz
gunicorn
